versions/termux-version/requirements.txt
fastapi
uvicorn
jinja2
httpx
pydantic
# Mobile-friendly LLM runtimes for Termux
llama-cpp-python
onnxruntime
# accelerate (optional, but keep for small CPU optimizations)
accelerate
